+++ Model/Api/ShippingMethods.php
@@ -300,7 +300,7 @@ class ShippingMethods implements ShippingMethodsInterface
     public function getShippingMethods($cart, $shipping_address)
     {
         try {
-            // $this->logHelper->addInfoLog($this->request->getContent());
+//            $this->logHelper->addInfoLog($this->request->getContent());
 
             $this->hookHelper->setCommonMetaData();
             $this->hookHelper->setHeaders();
@@ -326,8 +326,13 @@ class ShippingMethods implements ShippingMethodsInterface
             $this->sessionHelper->loadSession($quote);
 
             $addressData = $this->cartHelper->handleSpecialAddressCases($shipping_address);
+
             $this->applyDiscountIfNotUsaShipping($quote, $addressData);
-            $this->validateAddressData($addressData);
+
+            if (isset($addressData['email']) && $addressData['email'] !== null) {
+                $this->validateAddressData($addressData);
+            }
+            
             $shippingOptionsModel = $this->shippingEstimation($quote, $addressData);
 
             if ($this->taxAdjusted) {