+++ CHANGELOG.md
@@ -13,3 +13,18 @@
 ## [v1.1.2] 2018-10-30
 ## [v1.1.3] 2018-11-27
 ## [v1.1.4] 2018-12-04
+## [v1.1.5](https://github.com/BoltApp/bolt-magento2/releases/tag/1.1.5) 2018-12-11
+ - Use circleCI instead of TravisCI
+ - Prevent order ceation API call with an empty cart
+ - Complete order stays in payment review state on a long hook delay fix
+ - Invalid capture amount failed hook fix
+## [v1.1.6](https://github.com/BoltApp/bolt-magento2/releases/tag/1.1.6) 2018-12-13
+ - Force approve/reject failed hook fix
+## [v1.1.7](https://github.com/BoltApp/bolt-magento2/releases/tag/1.1.7) 2018-12-21
+ - Amasty Gift Card support
+ - No status after unhold fix
+## [v1.1.8](https://github.com/BoltApp/bolt-magento2/releases/tag/1.1.8) 2019-01-09
+ - Check if order payment method is 'boltpay'
+ - Add currency_code field to cart currency data
+ - Dispatch sales_quote_save_after event for active (parent) quotes only
+ - Fixed consistency for Amasty Gift Card module